Bug Fix and optimisation
