+++ alps-ios-app-color.xcodeproj/xcshareddata/xcschemes/alps-ios-app-color.xcscheme
@@ -83,6 +83,13 @@
             ReferencedContainer = "container:alps-ios-app-color.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
+      <EnvironmentVariables>
+         <EnvironmentVariable
+            key = "OS_ACTIVITY_MODE"
+            value = "disable"
+            isEnabled = "YES">
+         </EnvironmentVariable>
+      </EnvironmentVariables>
       <AdditionalOptions>
       </AdditionalOptions>
    </LaunchAction>
