+++ alps-ios-app-color.xcodeproj/xcuserdata/jmd.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,7 +4,7 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
-		<key>alps-ios-app-color.xcscheme</key>
+		<key>alps-ios-app-color.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>6</integer>
